Obtaining the cutpoints and / or regroupments of a discretization.

This defines the generic method "cutpoints" which will provide the cutpoints of a discretization scheme of S4 class glmdisc.

This defines the method to provide the cutpoints of a trained glmdisc.

Arguments

glmdisc
The trained glmdisc S4 object.

Examples

# Simulation of a discretized logit model
set.seed(1)
x = matrix(runif(300), nrow = 100, ncol = 3)
cuts = seq(0,1,length.out= 4)
xd = apply(x,2, function(col) as.numeric(cut(col,cuts)))
theta = t(matrix(c(0,0,0,2,2,2,-2,-2,-2),ncol=3,nrow=3))
log_odd = rowSums(t(sapply(seq_along(xd[,1]), function(row_id) sapply(seq_along(xd[row_id,]),
function(element) theta[xd[row_id,element],element]))))
y = rbinom(100,1,1/(1+exp(-log_odd)))

sem_disc <- glmdisc(x,y,iter=50,m_start=4,test=FALSE,validation=FALSE,criterion="aic")

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.04880875810219e+32

#> Iteration 1 ended with a performance of aic = -Inf

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.2089883658084e-09

#> Iteration 2 ended with a performance of aic = -133.949174420269

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.90836103177297e-10

#> Iteration 3 ended with a performance of aic = -131.125366036395

#> Current interaction being tested is between variables  and  with a probability of acceptance of 8.08816886713127e-10

#> Iteration 4 ended with a performance of aic = -130.664923475645

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.90884131247991e-14

#> Iteration 5 ended with a performance of aic = -131.352006941716

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.79079792587077e-14

#> Iteration 6 ended with a performance of aic = -103.774699780022

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.17062784632716e-24

#> Iteration 7 ended with a performance of aic = -103.41780377928

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.14436535339994e-13

#> Iteration 8 ended with a performance of aic = -99.0128549467669

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.53877878344639e-18

#> Iteration 9 ended with a performance of aic = -98.9821294699303

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.44086692471949e-17

#> Iteration 10 ended with a performance of aic = -100.152569451479

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.38207783674365e-19

#> Iteration 11 ended with a performance of aic = -89.7049122036646

#> Current interaction being tested is between variables  and  with a probability of acceptance of 6.98178703314176e-16

#> Iteration 12 ended with a performance of aic = -97.3768061662805

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.75224658553544e-19

#> Iteration 13 ended with a performance of aic = -109.91011816519

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.4135328603722e-13

#> Iteration 14 ended with a performance of aic = -91.5251682041485

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.96858140477279e-18

#> Iteration 15 ended with a performance of aic = -98.3738636837189

#> Current interaction being tested is between variables  and  with a probability of acceptance of 5.13081374698454e-13

#> Iteration 16 ended with a performance of aic = -102.577035391326

#> Current interaction being tested is between variables  and  with a probability of acceptance of 5.15014248455162e-11

#> Iteration 17 ended with a performance of aic = -113.551366305511

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.5738541285715e-14

#> Iteration 18 ended with a performance of aic = -104.961051412249

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.24686224255693e-20

#> Iteration 19 ended with a performance of aic = -110.471512237769

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.45088418762174e-17

#> Iteration 20 ended with a performance of aic = -115.332378497203

#> Current interaction being tested is between variables  and  with a probability of acceptance of 5.23668214705287e-18

#> Iteration 21 ended with a performance of aic = -102.778622889863

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.15658157461585e-12

#> Iteration 22 ended with a performance of aic = -112.344644485233

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.47387532952809e-19

#> Iteration 23 ended with a performance of aic = -115.119226164277

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.02044454281748e-12

#> Iteration 24 ended with a performance of aic = -114.36366944141

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.00782015300449e-18

#> Iteration 25 ended with a performance of aic = -100.364276988216

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.0455999966163e-17

#> Iteration 26 ended with a performance of aic = -92.5505730694462

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.31696359319126e-13

#> Iteration 27 ended with a performance of aic = -90.5820198565694

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.21980470550277e-19

#> Iteration 28 ended with a performance of aic = -104.394592676889

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.26683826906696e-13

#> Iteration 29 ended with a performance of aic = -95.4483558370767

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.17867986517599e-14

#> Iteration 30 ended with a performance of aic = -101.371115247309

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.012130310518534

#> Iteration 31 ended with a performance of aic = -100.037502306596

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.0790818417476765

#> Iteration 32 ended with a performance of aic = -99.1856100146779

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.0790818417476765

#> Iteration 33 ended with a performance of aic = -100.136420158839

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.91139287822405e-20

#> Iteration 34 ended with a performance of aic = -102.383020881871

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.31425530196366e-20

#> Iteration 35 ended with a performance of aic = -95.3165619375518

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.0790818417476765

#> Iteration 36 ended with a performance of aic = -92.8152535266965

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.11587839157045e-13

#> Iteration 37 ended with a performance of aic = -96.9640038415482

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.65801885371022e-09

#> Iteration 38 ended with a performance of aic = -100.620586128701

#> Current interaction being tested is between variables  and  with a probability of acceptance of 8.43469018548795e-10

#> Iteration 39 ended with a performance of aic = -96.8356781448577

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.26170544834111e-10

#> Iteration 40 ended with a performance of aic = -92.2641510333789

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.0790818417476765

#> Iteration 41 ended with a performance of aic = -92.5389673408644

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.24143299616622e-10

#> Iteration 42 ended with a performance of aic = -92.2493514501428

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.0790818417476765

#> Iteration 43 ended with a performance of aic = -94.3262588724134

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.0790818417476765

#> Iteration 44 ended with a performance of aic = -107.456649395821

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.0790818417476765

#> Iteration 45 ended with a performance of aic = -106.247232289959

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.0790818417476765

#> Iteration 46 ended with a performance of aic = -102.871341028718

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.28877418269782e-09

#> Iteration 47 ended with a performance of aic = -99.4926880744398

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.32132599704629e-10

#> Iteration 48 ended with a performance of aic = -115.088340126686

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.90852165804767e-10

#> Iteration 49 ended with a performance of aic = -102.082587221016

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.0790818417476765

#> Iteration 50 ended with a performance of aic = -109.306140101275

cutpoints(sem_disc)

#> $X1
#> [1] 0.2258860 0.6614014
#> 
#> $X2
#> [1] 0.1608269 0.7364356
#> 
#> $X3
#> [1] 0.08181332 0.54110570
#>