glmdisc: A package for discretizing continuous features, grouping categorical features' values and optimizing it for logistic regression.

The glmdisc package provides two important functions: glmdisc and its associate method discretize.

glmdisc function

The glmdisc function discretizes a training set using an SEM-Gibbs based method.

discretize function

The discretize function will discretize a new input dataset given a discretization scheme of S4 class glmdisc.

cutpoints function

The cutpoints function will provide the cutpoints / groupings of a discretization scheme of S4 class glmdisc in a list.

predict function

The predict function will discretize a raw test set, given a provided discretization scheme of S4 class glmdisc, using the discretize function and return the predicted probabilities.

Miscellaneous

We provide as well the classical show, print, summary functions, as well as normalizedGini that is used to calculate the Gini index (a classical Credit Scoring goodness-of-fit indicator).