Obtaining the cutpoints and / or regroupments of a discretization.

This defines the generic method "cutpoints" which will provide the cutpoints of a discretization scheme of S4 class glmdisc.

This defines the method to provide the cutpoints of a trained glmdisc.

cutpoints(object)

# S4 method for glmdisc
cutpoints(object)

Arguments

object
generic glmdisc object

glmdisc
The trained glmdisc S4 object.

Examples

# Simulation of a discretized logit model
set.seed(1)
x <- matrix(runif(300), nrow = 100, ncol = 3)
cuts <- seq(0, 1, length.out = 4)
xd <- apply(x, 2, function(col) as.numeric(cut(col, cuts)))
theta <- t(matrix(c(0, 0, 0, 2, 2, 2, -2, -2, -2), ncol = 3, nrow = 3))
log_odd <- rowSums(t(sapply(seq_along(xd[, 1]), function(row_id) {
  sapply(
    seq_along(xd[row_id, ]),
    function(element) theta[xd[row_id, element], element]
  )
})))
y <- rbinom(100, 1, 1 / (1 + exp(-log_odd)))

sem_disc <- glmdisc(x, y,
  iter = 50, m_start = 4, test = FALSE,
  validation = FALSE, criterion = "aic"
)

#> Current interaction being tested is between variables  and  with a probability of acceptance of 8.60794054684477e+22

#> Iteration 1 ended with a performance of aic = -Inf

#> Current interaction being tested is between variables  and  with a probability of acceptance of 454070997000659

#> Iteration 2 ended with a performance of aic = -146.627387885044

#> Current interaction being tested is between variables  and  with a probability of acceptance of 39859840.420523

#> Iteration 3 ended with a performance of aic = -127.764828492265

#> Current interaction being tested is between variables  and  with a probability of acceptance of 8.31209440743563e-10

#> Iteration 4 ended with a performance of aic = -127.009783763371

#> Current interaction being tested is between variables  and  with a probability of acceptance of 5.31771086618256e-09

#> Iteration 5 ended with a performance of aic = -125.336379885988

#> Current interaction being tested is between variables  and  with a probability of acceptance of 5.90823043287905e-13

#> Iteration 6 ended with a performance of aic = -118.694287498429

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.06112864710227e-09

#> Iteration 7 ended with a performance of aic = -105.140665981986

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.17930959880837e-11

#> Iteration 8 ended with a performance of aic = -117.521689451285

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.03495038363605e-17

#> Iteration 9 ended with a performance of aic = -96.0062145758655

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.87392457900451e-18

#> Iteration 10 ended with a performance of aic = -123.222853601946

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.94627926349477e-06

#> Iteration 11 ended with a performance of aic = -116.180471015991

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.19124696405324e-18

#> Iteration 12 ended with a performance of aic = -113.131586033604

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.35079776783696e-14

#> Iteration 13 ended with a performance of aic = -94.6628211447204

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.90818482108363e-14

#> Iteration 14 ended with a performance of aic = -101.86035338346

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.91604343743075e-10

#> Iteration 15 ended with a performance of aic = -107.530239359646

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.15130098621135e-17

#> Iteration 16 ended with a performance of aic = -90.3286583488307

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.41554685597652e-17

#> Iteration 17 ended with a performance of aic = -95.8589726784055

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.11459160387099e-18

#> Iteration 18 ended with a performance of aic = -94.400027206177

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.90825613536406e-10

#> Iteration 19 ended with a performance of aic = -100.662210428279

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.70913316675428e-14

#> Iteration 20 ended with a performance of aic = -97.5262609601426

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.90817688741452e-18

#> Iteration 21 ended with a performance of aic = -97.0341401684959

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.52401724347715e-17

#> Iteration 22 ended with a performance of aic = -99.7807932827513

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.90817625115562e-14

#> Iteration 23 ended with a performance of aic = -99.6819123188817

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.90815594064724e-10

#> Iteration 24 ended with a performance of aic = -105.112947788497

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.04003116575161e-13

#> Iteration 25 ended with a performance of aic = -93.698414773591

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.51867530443759e-13

#> Iteration 26 ended with a performance of aic = -98.5239377655172

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.90818208680709e-14

#> Iteration 27 ended with a performance of aic = -102.015750281122

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.54405276151252e-13

#> Iteration 28 ended with a performance of aic = -103.402940956239

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.44849443317806e-09

#> Iteration 29 ended with a performance of aic = -96.9711288214307

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.95506822655655e-09

#> Iteration 30 ended with a performance of aic = -115.057137799568

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.5442087848914e-10

#> Iteration 31 ended with a performance of aic = -104.982509167778

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.40190318641173e-12

#> Iteration 32 ended with a performance of aic = -89.2971105270865

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.64928594304378e-16

#> Iteration 33 ended with a performance of aic = -96.8513820687929

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.18224332822318e-17

#> Iteration 34 ended with a performance of aic = -104.539586619264

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.03132037788868e-19

#> Iteration 35 ended with a performance of aic = -91.7186652471138

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.72127263028884e-20

#> Iteration 36 ended with a performance of aic = -108.992482391601

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.21384488130027e-16

#> Iteration 37 ended with a performance of aic = -113.598054832842

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.36804185413001e-16

#> Iteration 38 ended with a performance of aic = -115.461466644513

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.5171308145283e-25

#> Iteration 39 ended with a performance of aic = -122.475176571153

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.30945378595437e-22

#> Iteration 40 ended with a performance of aic = -114.997597975846

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.51338954030697e-23

#> Iteration 41 ended with a performance of aic = -123.539974449971

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.34569537612638e-22

#> Iteration 42 ended with a performance of aic = -113.56101777117

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.49339067119636e-24

#> Iteration 43 ended with a performance of aic = -107.991169862016

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.96643280395445e-24

#> Iteration 44 ended with a performance of aic = -104.054061363204

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.75125072278561e-19

#> Iteration 45 ended with a performance of aic = -99.2612311815615

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.9200205579114e-25

#> Iteration 46 ended with a performance of aic = -89.3974217247178

#> Current interaction being tested is between variables  and  with a probability of acceptance of 6.25189161692065e-25

#> Iteration 47 ended with a performance of aic = -96.8091312349041

#> Current interaction being tested is between variables  and  with a probability of acceptance of 5.28863572926684e-17

#> Iteration 48 ended with a performance of aic = -110.829497011994

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.34138893832509e-25

#> Iteration 49 ended with a performance of aic = -108.821660279839

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.36408599632033e-23

#> Iteration 50 ended with a performance of aic = -102.857294908262

cutpoints(sem_disc)

#> $X1
#> [1] 0.1148178 0.6489653
#> 
#> $X2
#> [1] 0.697266
#> 
#> $X3
#> [1] 0.5411057 0.6404998
#>