Model-based multivariate discretization for logistic regression.

This function discretizes a training set using an SEM-Gibbs based method (see References section). It detects numerical features of the dataset and discretizes them ; values of categorical features (of type factor) are regrouped. This is done in a multivariate supervised way. Assessment of the correct model is done via AIC, BIC or test set error (see parameter criterion). Second-order interactions can be searched through the optional interaction parameter using a Metropolis-Hastings algorithm (see References section).

glmdisc(predictors, labels, interact = TRUE, validation = TRUE,
  test = TRUE, criterion = "gini", iter = 1000, m_start = 20,
  reg_type = "poly", proportions = c(0.2, 0.2))

Arguments

predictors
The matrix array containing the numerical or factor attributes to discretize.

labels
The actual labels of the provided predictors (0/1).

interact
Boolean : True (default) if interaction detection is wanted (Warning: may be very memory/time-consuming).

validation
Boolean : True if the algorithm should use predictors to construct a validation set on which to search for the best discretization scheme using the provided criterion (default: TRUE).

test
Boolean : True if the algorithm should use predictors to construct a test set on which to calculate the provided criterion using the best discretization scheme (chosen thanks to the provided criterion on either the test set (if true) or the training set (otherwise)) (default: TRUE).

criterion
The criterion ('gini','aic','bic') to use to choose the best discretization scheme among the generated ones (default: 'gini'). Nota Bene: it is best to use 'gini' only when test is set to TRUE and 'aic' or 'bic' when it is not. When using 'aic' or 'bic' with a test set, the likelihood is returned as there is no need to penalize for generalization purposes.

iter
The number of iterations to do in the SEM protocole (default: 1000).

m_start
The maximum number of resulting categories for each variable wanted (default: 20).

reg_type
The model to use between discretized and continuous features (currently, only multinomial logistic regression ('poly') and ordered logistic regression ('polr') are supported ; default: 'poly'). WARNING: 'poly' requires the mnlogit package, 'polr' requires the MASS package.

proportions
The list of the proportions wanted for test and validation set. Not used when both test and validation are false. Only the first is used when there is only one of either test or validation that is set to TRUE. Produces an error when the sum is greater to one. Default: list(0.2,0.2) so that the training set has 0.6 of the input observations.

Details

This function finds the most appropriate discretization scheme for logistic regression. When provided with a continuous variable \(X\), it tries to convert it to a categorical variable \(E\) which values uniquely correspond to intervals of the continuous variable \(X\). When provided with a categorical variable \(X\), it tries to find the best regroupement of its values and subsequently creates categorical variable \(E\). The goal is to perform supervised learning with logistic regression so that you have to specify a target variable \(Y\) denoted by labels. The ‘‘discretization'' process, i.e. the transformation of \(X\) to \(E\) is done so as to achieve the best logistic regression model \(p(y|e;\theta)\). It can be interpreted as a special case feature engineering algorithm. Subsequently, its outputs are: the optimal discretization scheme and the logistic regression model associated with it. We also provide the parameters that were provided to the function and the evolution of the criterion with respect to the algorithm's iterations.

References

Celeux, G., Chauveau, D., Diebolt, J. (1995), On Stochastic Versions of the EM Algorithm. [Research Report] RR-2514, INRIA. 1995. <inria-00074164>

Agresti, A. (2002) Categorical Data. Second edition. Wiley.

See also

Examples

# Simulation of a discretized logit model
set.seed(1)
x = matrix(runif(300), nrow = 100, ncol = 3)
cuts = seq(0,1,length.out= 4)
xd = apply(x,2, function(col) as.numeric(cut(col,cuts)))
theta = t(matrix(c(0,0,0,2,2,2,-2,-2,-2),ncol=3,nrow=3))
log_odd = rowSums(t(sapply(seq_along(xd[,1]), function(row_id) sapply(seq_along(xd[row_id,]),
function(element) theta[xd[row_id,element],element]))))
y = rbinom(100,1,1/(1+exp(-log_odd)))

sem_disc <- glmdisc(x,y,iter=50,m_start=4,test=FALSE,validation=FALSE,criterion="aic")

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2049222767148599296

#> Iteration 1 ended with a performance of aic = -154.841857064322

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.00420897514494496

#> Iteration 2 ended with a performance of aic = -129.949175662963

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.000790819416889533

#> Iteration 3 ended with a performance of aic = -117.436861179737

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.04160399081207e-05

#> Iteration 4 ended with a performance of aic = -104.831846812578

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.12127092523626e-05

#> Iteration 5 ended with a performance of aic = -106.481666238122

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.27370629942661e-05

#> Iteration 6 ended with a performance of aic = -126.452567349744

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.09666409683387e-12

#> Iteration 7 ended with a performance of aic = -116.08812043538

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.36013340045633e-05

#> Iteration 8 ended with a performance of aic = -112.796907496068

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.4143373309866e-05

#> Iteration 9 ended with a performance of aic = -110.112221042731

#> Current interaction being tested is between variables  and  with a probability of acceptance of 6.33824469348222e-08

#> Iteration 10 ended with a performance of aic = -105.91805804688

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.26014504983102e-09

#> Iteration 11 ended with a performance of aic = -106.418311256461

#> Current interaction being tested is between variables  and  with a probability of acceptance of 6.25411868185495e-06

#> Iteration 12 ended with a performance of aic = -110.007586256327

#> Current interaction being tested is between variables  and  with a probability of acceptance of 0.00447280124530351

#> Iteration 13 ended with a performance of aic = -103.440099780966

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.03875591821536e-05

#> Iteration 14 ended with a performance of aic = -97.7089238043403

#> Current interaction being tested is between variables  and  with a probability of acceptance of 6.3296439178279e-09

#> Iteration 15 ended with a performance of aic = -96.60307808566

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.08276578408124e-14

#> Iteration 16 ended with a performance of aic = -94.6103393512782

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.19580200127219e-06

#> Iteration 17 ended with a performance of aic = -96.0106144384931

#> Current interaction being tested is between variables  and  with a probability of acceptance of 6.96992608047388e-12

#> Iteration 18 ended with a performance of aic = -90.3174647902317

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.0493459409431e-13

#> Iteration 19 ended with a performance of aic = -89.9215774815297

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.53728501242877e-14

#> Iteration 20 ended with a performance of aic = -95.9874051221171

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.24620267567388e-08

#> Iteration 21 ended with a performance of aic = -94.1555377283248

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.04527596670886e-06

#> Iteration 22 ended with a performance of aic = -89.2083183862152

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.46269496415357e-12

#> Iteration 23 ended with a performance of aic = -90.3013474340092

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.1428578600483e-11

#> Iteration 24 ended with a performance of aic = -87.4721211032281

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.36878986313949e-07

#> Iteration 25 ended with a performance of aic = -84.7159337945391

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.01978212375338e-06

#> Iteration 26 ended with a performance of aic = -86.5042258731647

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.08238682658476e-12

#> Iteration 27 ended with a performance of aic = -91.0081467344218

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.69037074587262e-11

#> Iteration 28 ended with a performance of aic = -93.3732243103678

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.78924488734201e-11

#> Iteration 29 ended with a performance of aic = -91.5376211693954

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.09170090928787e-11

#> Iteration 30 ended with a performance of aic = -92.9473389204547

#> Current interaction being tested is between variables  and  with a probability of acceptance of 3.89961802880441e-09

#> Iteration 31 ended with a performance of aic = -89.6521410680045

#> Current interaction being tested is between variables  and  with a probability of acceptance of 6.75506211325842e-09

#> Iteration 32 ended with a performance of aic = -78.0915498296754

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.99774828212841e-05

#> Iteration 33 ended with a performance of aic = -76.8167844149011

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.72789642568013e-05

#> Iteration 34 ended with a performance of aic = -80.2215819834547

#> Current interaction being tested is between variables  and  with a probability of acceptance of 7.69142891126253e-11

#> Iteration 35 ended with a performance of aic = -83.3090051775009

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.00550037564027e-08

#> Iteration 36 ended with a performance of aic = -83.4752551238383

#> Current interaction being tested is between variables  and  with a probability of acceptance of 8.20228743211535e-07

#> Iteration 37 ended with a performance of aic = -79.7463052397814

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.08620793899292e-08

#> Iteration 38 ended with a performance of aic = -80.4466017835961

#> Current interaction being tested is between variables  and  with a probability of acceptance of 2.069235335506e-08

#> Iteration 39 ended with a performance of aic = -81.6771892628792

#> Current interaction being tested is between variables  and  with a probability of acceptance of 5.45867914531335e-11

#> Iteration 40 ended with a performance of aic = -81.4387248622089

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.71151984445067e-08

#> Iteration 41 ended with a performance of aic = -85.1005832819567

#> Current interaction being tested is between variables  and  with a probability of acceptance of 4.01650969333901e-08

#> Iteration 42 ended with a performance of aic = -82.8917066701289

#> Current interaction being tested is between variables  and  with a probability of acceptance of 5.46187020797332e-09

#> Iteration 43 ended with a performance of aic = -79.9793086991862

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.26672766180949e-08

#> Iteration 44 ended with a performance of aic = -83.16599065534

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.33328817882801e-09

#> Iteration 45 ended with a performance of aic = -81.0394779696091

#> Current interaction being tested is between variables  and  with a probability of acceptance of 6.80128509754935e-09

#> Iteration 46 ended with a performance of aic = -84.5603842694265

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.46270984048275e-09

#> Iteration 47 ended with a performance of aic = -78.2320193252781

#> Current interaction being tested is between variables  and  with a probability of acceptance of 1.90706664694385e-07

#> Iteration 48 ended with a performance of aic = -74.119824440545

#> Current interaction being tested is between variables  and  with a probability of acceptance of 9.3538893896028e-09

#> Iteration 49 ended with a performance of aic = -79.6766040003919

#> Current interaction being tested is between variables  and  with a probability of acceptance of 5.67166143991618e-09

#> Iteration 50 ended with a performance of aic = -76.6148187907212

discretize(sem_disc,data.frame(x))

#>        [,1] [,2] [,3]
#>   [1,] "1"  "4"  "3" 
#>   [2,] "4"  "4"  "3" 
#>   [3,] "2"  "3"  "4" 
#>   [4,] "3"  "2"  "3" 
#>   [5,] "1"  "4"  "3" 
#>   [6,] "3"  "3"  "4" 
#>   [7,] "3"  "3"  "4" 
#>   [8,] "2"  "4"  "3" 
#>   [9,] "2"  "2"  "3" 
#>  [10,] "1"  "4"  "1" 
#>  [11,] "1"  "2"  "1" 
#>  [12,] "1"  "2"  "3" 
#>  [13,] "3"  "4"  "1" 
#>  [14,] "4"  "4"  "1" 
#>  [15,] "3"  "3"  "1" 
#>  [16,] "2"  "3"  "4" 
#>  [17,] "3"  "2"  "1" 
#>  [18,] "3"  "3"  "1" 
#>  [19,] "4"  "4"  "1" 
#>  [20,] "3"  "4"  "4" 
#>  [21,] "3"  "2"  "3" 
#>  [22,] "1"  "4"  "3" 
#>  [23,] "2"  "4"  "4" 
#>  [24,] "1"  "3"  "4" 
#>  [25,] "1"  "2"  "1" 
#>  [26,] "4"  "4"  "4" 
#>  [27,] "1"  "4"  "3" 
#>  [28,] "4"  "3"  "3" 
#>  [29,] "3"  "3"  "4" 
#>  [30,] "4"  "4"  "1" 
#>  [31,] "2"  "4"  "4" 
#>  [32,] "2"  "3"  "3" 
#>  [33,] "2"  "3"  "4" 
#>  [34,] "1"  "4"  "1" 
#>  [35,] "3"  "2"  "4" 
#>  [36,] "3"  "4"  "1" 
#>  [37,] "3"  "4"  "1" 
#>  [38,] "1"  "4"  "4" 
#>  [39,] "3"  "2"  "4" 
#>  [40,] "2"  "4"  "4" 
#>  [41,] "3"  "4"  "4" 
#>  [42,] "2"  "4"  "4" 
#>  [43,] "3"  "3"  "1" 
#>  [44,] "2"  "3"  "3" 
#>  [45,] "2"  "2"  "4" 
#>  [46,] "3"  "4"  "3" 
#>  [47,] "1"  "3"  "4" 
#>  [48,] "2"  "2"  "3" 
#>  [49,] "3"  "3"  "4" 
#>  [50,] "3"  "2"  "1" 
#>  [51,] "2"  "4"  "1" 
#>  [52,] "3"  "4"  "1" 
#>  [53,] "2"  "4"  "4" 
#>  [54,] "1"  "4"  "1" 
#>  [55,] "1"  "4"  "4" 
#>  [56,] "1"  "3"  "3" 
#>  [57,] "4"  "4"  "3" 
#>  [58,] "2"  "3"  "4" 
#>  [59,] "2"  "3"  "4" 
#>  [60,] "2"  "3"  "1" 
#>  [61,] "3"  "3"  "4" 
#>  [62,] "1"  "2"  "4" 
#>  [63,] "2"  "4"  "3" 
#>  [64,] "4"  "2"  "1" 
#>  [65,] "2"  "2"  "1" 
#>  [66,] "1"  "4"  "3" 
#>  [67,] "2"  "3"  "3" 
#>  [68,] "3"  "4"  "1" 
#>  [69,] "1"  "2"  "1" 
#>  [70,] "3"  "4"  "3" 
#>  [71,] "4"  "4"  "3" 
#>  [72,] "3"  "2"  "3" 
#>  [73,] "4"  "2"  "4" 
#>  [74,] "4"  "4"  "3" 
#>  [75,] "2"  "4"  "3" 
#>  [76,] "3"  "2"  "3" 
#>  [77,] "3"  "4"  "3" 
#>  [78,] "4"  "2"  "3" 
#>  [79,] "3"  "4"  "4" 
#>  [80,] "3"  "2"  "1" 
#>  [81,] "2"  "4"  "3" 
#>  [82,] "3"  "3"  "4" 
#>  [83,] "4"  "2"  "1" 
#>  [84,] "4"  "4"  "4" 
#>  [85,] "3"  "2"  "3" 
#>  [86,] "1"  "3"  "3" 
#>  [87,] "3"  "2"  "4" 
#>  [88,] "1"  "2"  "3" 
#>  [89,] "1"  "2"  "3" 
#>  [90,] "1"  "4"  "3" 
#>  [91,] "1"  "2"  "3" 
#>  [92,] "1"  "4"  "3" 
#>  [93,] "2"  "3"  "1" 
#>  [94,] "3"  "2"  "4" 
#>  [95,] "3"  "3"  "4" 
#>  [96,] "3"  "4"  "1" 
#>  [97,] "2"  "3"  "3" 
#>  [98,] "2"  "2"  "3" 
#>  [99,] "3"  "4"  "3" 
#> [100,] "2"  "2"  "1"